Add refile controls to HierarchyWidget
